FIX: Do not verify the `iat` claim in JWT tokens
The JWT specification (https://tools.ietf.org/html/rfc7519#page-10) does not require verification of this claim. If the issuer wishes to restrict the validity of the token, they can use the 'nbf' (not before) claim which is intended for this purpose. Discourse will verify the `nbf` claim if it is present.

In practice, clock skew between identity providers and Discourse was causing JWT validity errors to be raised.